+++ nilearn/tests/test_param_validation.py
@@ -3,13 +3,23 @@ Test the _utils.param_validation module
 """
 
 import numpy as np
+import warnings
+import os
+import nibabel
 
-from nose.tools import assert_true, assert_equal
+from nose.tools import assert_equal, assert_true, assert_raises
+from sklearn.base import BaseEstimator
 
 from nilearn._utils.testing import assert_raises_regex, assert_warns
 
 from nilearn._utils.extmath import fast_abs_percentile
-from nilearn._utils.param_validation import check_threshold
+from nilearn._utils.param_validation import (MNI152_BRAIN_VOLUME,
+                                             _get_mask_volume,
+                                             check_feature_screening,
+                                             check_threshold)
+
+mni152_brain_mask = (
+    "/usr/share/fsl/data/standard/MNI152_T1_1mm_brain_mask.nii.gz")
 
 
 def test_check_threshold():
@@ -54,3 +64,35 @@ def test_check_threshold():
     assert_true(1. < check_threshold("50%", matrix,
                                      percentile_func=fast_abs_percentile,
                                      name=name) <= 2.)
+
+
+def test_feature_screening():
+    for is_classif in [True, False]:
+        for screening_percentile in [100, None, 20, 101, -1, 10]:
+
+            if screening_percentile == 100 or screening_percentile is None:
+                assert_equal(check_feature_screening(
+                    screening_percentile, MNI152_BRAIN_VOLUME, is_classif),
+                    None)
+            elif screening_percentile == 101 or screening_percentile == -1:
+                assert_raises(ValueError, check_feature_screening,
+                              screening_percentile, MNI152_BRAIN_VOLUME,
+                              is_classif)
+            elif screening_percentile == 20:
+                assert_true(isinstance(check_feature_screening(
+                    screening_percentile, MNI152_BRAIN_VOLUME, is_classif),
+                    BaseEstimator))
+            else:
+                assert_warns(UserWarning, check_feature_screening,
+                             screening_percentile, MNI152_BRAIN_VOLUME * 2,
+                             is_classif)
+
+
+def test_get_mask_volume():
+    # Test that hard-coded standard mask volume can be corrected computed
+    if os.path.isfile(mni152_brain_mask):
+        assert_equal(MNI152_BRAIN_VOLUME, _get_mask_volume(nibabel.load(
+            mni152_brain_mask)))
+    else:
+        warnings.warn("Couldn't find %s (for testing)" % (mni152_brain_mask))
+
